blog-webappaad-backend/requirements.txt
azure-functions
python-jose>=3.0.0
pyodbc